Experiment 7 Face Recognition/requirements.txt
dlib-bin
numpy
pandas
Pillow
scikit-learn
